third_party/WebKit/LayoutTests/svg/animations/getCurrentTime-pause-unpause-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS svg.getCurrentTime() is 0
PASS svg.getCurrentTime() is 0.05
PASS svg.getCurrentTime() is 0.05
